+++ src/main/java/com/smunity/api/domain/account/dto/SignUpDto.java
@@ -1,9 +1,9 @@
 package com.smunity.api.domain.account.dto;
 
-import com.smunity.api.domain.account.domain.Department;
-import com.smunity.api.domain.account.domain.Profile;
-import com.smunity.api.domain.account.domain.User;
-import com.smunity.api.domain.account.domain.Year;
+import com.smunity.api.domain.account.entity.Department;
+import com.smunity.api.domain.account.entity.Profile;
+import com.smunity.api.domain.account.entity.User;
+import com.smunity.api.domain.account.entity.Year;
 import lombok.Data;
