+++ src/main/java/com/smunity/api/domain/account/repository/DepartmentRepository.java
@@ -1,6 +1,6 @@
 package com.smunity.api.domain.account.repository;
 
-import com.smunity.api.domain.account.domain.Department;
+import com.smunity.api.domain.account.entity.Department;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
