+++ src/main/java/com/smunity/api/domain/account/repository/YearRepository.java
@@ -1,6 +1,6 @@
 package com.smunity.api.domain.account.repository;
 
-import com.smunity.api.domain.account.domain.Year;
+import com.smunity.api.domain.account.entity.Year;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
