+++ src/main/java/com/smunity/api/domain/petition/service/impl/RespondServiceImpl.java
@@ -1,6 +1,6 @@
 package com.smunity.api.domain.petition.service.impl;
 
-import com.smunity.api.domain.account.domain.User;
+import com.smunity.api.domain.account.entity.User;
 import com.smunity.api.domain.account.repository.UserRepository;
 import com.smunity.api.domain.petition.entity.Respond;
 import com.smunity.api.domain.petition.entity.Petition;
