+++ src/main/java/com/smunity/api/domain/question/service/impl/QuestionServiceImpl.java
@@ -1,6 +1,6 @@
 package com.smunity.api.domain.question.service.impl;
 
-import com.smunity.api.domain.account.domain.User;
+import com.smunity.api.domain.account.entity.User;
 import com.smunity.api.domain.account.repository.UserRepository;
 import com.smunity.api.domain.question.domain.Question;
 import com.smunity.api.domain.question.dto.QuestionDto;
