+++ src/test/java/com/smunity/api/domain/account/repository/UserRepositoryTest.java
@@ -1,7 +1,6 @@
 package com.smunity.api.domain.account.repository;
 
-import com.smunity.api.domain.account.domain.User;
-import com.smunity.api.domain.account.repository.UserRepository;
+import com.smunity.api.domain.account.entity.User;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
