+++ src/test/java/com/smunity/api/domain/account/repository/YearRepositoryTest.java
@@ -1,6 +1,6 @@
 package com.smunity.api.domain.account.repository;
 
-import com.smunity.api.domain.account.domain.Year;
+import com.smunity.api.domain.account.entity.Year;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
